functionApp/EpisodesGen/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-storage-blob
elevenlabs==0.2.27
ibm_cloud_sdk_core==3.19.2
ibm_watson==8.0.0
langchain==0.1.13
moviepy==1.0.3
nltk==3.8.1
numpy==1.24.2
openai==0.28.0
pdfminer.six==20221105
pdfplumber==0.10.4
Pillow==10.2.0
pydub==0.25.1
pymongo==4.6.2
python-dotenv==1.0.1
Requests==2.31.0
textblob==0.18.0.post0
transformers==4.38.1
sentence-transformers
torch
faiss-cpu